requirements.txt
pyshark
netifaces
httpx
asyncio
requests
flask
python-dotenv
